requirements.txt
groq
AppOpener
keyboard
beautifulsoup4
googlesearch-python
pywhatkit
python-dotenv
